Simplify PermissionGrant
