+++ raphtory-graphql/tests/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod graphql;
 pub mod strategy;
+pub mod tree;
